feat: add production deployment configuration

+++ api/controllers/userController.js
@@ -1,10 +1,10 @@
-import Listing from '../models/listing.model.js';
-import User from '../models/user.model.js';
-import { errorHandler } from '../utils/error.js';
+import Listing from "../models/listing.js";
+import User from "../models/user.js";
+import { errorHandler } from "../utils/error.js";
 
 export const updateUser = async (req, res, next) => {
   if (req.user.id !== req.params.id)
-    return next(errorHandler(401, 'You can only update your own account'));
+    return next(errorHandler(401, "You can only update your own account"));
   try {
     if (req.body.password) {
       req.body.password = bcrypt.hasSync(req.body.password, 10);
@@ -32,12 +32,18 @@ export const updateUser = async (req, res, next) => {
 
 export const deleteUser = async (req, res, next) => {
   if (req.user.id !== req.params.id)
-    return next(errorHandler(401, 'You can only delete your own account'));
+    return next(errorHandler(401, "You can only delete your own account"));
   try {
+    const user = await User.findById(req.params.id);
+    if (!user) return next(errorHandler(404, "User not found"));
+
+    await Listing.deleteMany({ userRef: req.params.id });
+
     await User.findByIdAndDelete(req.params.id);
-    res.clearCookie('access_token');
+
+    res.clearCookie("access_token");
     res.status(200).json({
-      message: 'User has been deleted!'
+      message: "User and all associated listings have been deleted!"
     });
   } catch (err) {
     next(err);
@@ -45,26 +51,19 @@ export const deleteUser = async (req, res, next) => {
 };
 
 export const getUserListings = async (req, res, next) => {
-  if (req.user.id === req.params.id) {
-    try {
-      const listings = await Listing.find({ userRef: req.params.id });
-      res.status(200).json(listings);
-    } catch (err) {
-      next(err);
-    }
-  } else {
-    return next(errorHandler(401, 'You can only view your own listings'));
+  try {
+    const listings = await Listing.find({ userRef: req.params.id });
+    res.status(200).json(listings);
+  } catch (error) {
+    next(errorHandler(500, "Error getting user listings"));
   }
 };
 
 export const getUser = async (req, res, next) => {
   try {
     const user = await User.findById(req.params.id);
-
-    if (!user) return next(errorHandler(404, 'User not found'));
-
+    if (!user) return next(errorHandler(404, "User not found"));
     const { password: pass, ...rest } = user._doc;
-
     res.status(200).json(rest);
   } catch (error) {
     next(error);
